RR-750 - Refactor router request handlers into their own source files
